Validate state transitions for nested state machines
There are two places that effectively contain a nested, self-contained state
machine:

- A Map state's Iterator
- A Parallel state's Branches

Previously the validation allowed both references from the containing state
machine into a nested one, and the reverse. Both of these are now correctly
disallowed.

The implementation of the transition checks is now more precise about which
state-referencing fields it looks at. For this reason, it no longer uses
jsonpath, which mostly seemed to be useful for easily finding fields (eg:
"Next") at arbitrary depth in the definition.

Not ideal, but I'm skipping the transition checks when there are schema
validation errors. This is because the transition checks no longer use jsonpath,
and so are less forgiving to invalidly-shaped input. Input that did not validate
against the schema could be any data structure at all, and so it's hard to
protect against throwing exceptions in the new transition-check code, as it
assumes that the input is mostly shaped like a valid State Machine.